+++ R/workflowtools-package.R
@@ -12,8 +12,9 @@ utils::globalVariables(c(
 #' @importFrom dplyr bind_rows filter pull
 #' @importFrom fs dir_info dir_ls path_dir path_expand path_file path_norm path_rel
 #' @importFrom furrr future_map_dfr future_map2_dfr
-#' @importFrom googledrive drive_ls drive_mkdir drive_put
+#' @importFrom googledrive drive_download drive_ls drive_mkdir drive_put
 #' @importFrom httr config set_config
+#' @importFrom purrr map2_dfr walk
 #' @importFrom remotes parse_github_repo_spec
 #' @importFrom rprojroot find_root from_wd is_git_root is_rstudio_project
 #' @importFrom sessioninfo session_info